1. Enhance the logic to check users's permissions of storage service 2. Add a new storage type "GROUP_PROFILE"

+++ turms/src/main/java/im/turms/turms/service/storage/StorageService.java
@@ -6,6 +6,7 @@ import im.turms.common.exception.TurmsBusinessException;
 import im.turms.turms.plugin.StorageServiceProvider;
 import im.turms.turms.plugin.TurmsPluginManager;
 import im.turms.turms.property.TurmsProperties;
+import im.turms.turms.service.group.GroupMemberService;
 import im.turms.turms.service.message.MessageService;
 import org.springframework.stereotype.Service;
 import org.springframework.validation.annotation.Validated;
@@ -20,38 +21,25 @@ public class StorageService {
     private final StorageServiceProvider provider;
     private final TurmsProperties turmsProperties;
     private final MessageService messageService;
+    private final GroupMemberService groupMemberService;
 
-    public StorageService(TurmsPluginManager turmsPluginManager, TurmsProperties turmsProperties, MessageService messageService) {
+    public StorageService(TurmsPluginManager turmsPluginManager, TurmsProperties turmsProperties, MessageService messageService, GroupMemberService groupMemberService) {
         this.provider = turmsPluginManager.getStorageServiceProvider();
         this.turmsProperties = turmsProperties;
         this.messageService = messageService;
+        this.groupMemberService = groupMemberService;
     }
 
-    public Mono<String> queryPreSignedGetUrl(@NotNull Long requesterId, @NotNull ContentType contentType, @Nullable String keyStr, @Nullable Long keyNum) {
+    public Mono<String> queryPresignedGetUrl(@NotNull Long requesterId, @NotNull ContentType contentType, @Nullable String keyStr, @Nullable Long keyNum) {
         if (provider != null) {
-            switch (contentType) {
-                case ATTACHMENT:
-                    if (keyNum != null) {
-                        return messageService.isMessageSentToUser(keyNum, requesterId)
-                                .flatMap(isSentToUser -> {
-                                    if (isSentToUser) {
-                                        return provider.queryPresignedGetUrl(requesterId, contentType, keyStr, keyNum);
-                                    } else {
-                                        return messageService.isMessageSentByUser(keyNum, requesterId)
-                                                .flatMap(isSentByUser -> {
-                                                    if (isSentByUser) {
-                                                        return provider.queryPresignedGetUrl(requesterId, contentType, keyStr, keyNum);
-                                                    } else {
-                                                        return Mono.error(TurmsBusinessException.get(TurmsStatusCode.UNAUTHORIZED));
-                                                    }
-                                                });
-                                    }
-                                });
-                    } else {
-                        throw TurmsBusinessException.get(TurmsStatusCode.ILLEGAL_ARGUMENTS);
-                    }
-            }
-            return provider.queryPresignedGetUrl(requesterId, contentType, keyStr, keyNum);
+            return hasPermissionToGet(requesterId, contentType, keyStr, keyNum)
+                    .flatMap(hasPermission -> {
+                        if (hasPermission) {
+                            return provider.queryPresignedGetUrl(requesterId, contentType, keyStr, keyNum);
+                        } else {
+                            return Mono.error(TurmsBusinessException.get(TurmsStatusCode.UNAUTHORIZED));
+                        }
+                    });
         } else {
             throw TurmsBusinessException.get(TurmsStatusCode.NOT_IMPLEMENTED);
         }
@@ -64,6 +52,9 @@ public class StorageService {
                 case PROFILE:
                     sizeLimit = turmsProperties.getStorage().getProfileSizeLimit();
                     break;
+                case GROUP_PROFILE:
+                    sizeLimit = turmsProperties.getStorage().getGroupProfileSizeLimit();
+                    break;
                 case ATTACHMENT:
                     sizeLimit = turmsProperties.getStorage().getAttachmentSizeLimit();
                     break;
@@ -71,7 +62,14 @@ public class StorageService {
                     throw new IllegalStateException("Unexpected value: " + contentType);
             }
             if (sizeLimit == 0 || contentLength <= sizeLimit) {
-                return provider.queryPresignedPutUrl(requesterId, contentType, keyStr, keyNum, contentLength);
+                return hasPermissionToPut(requesterId, contentType, keyStr, keyNum)
+                        .flatMap(hasPermission -> {
+                            if (hasPermission) {
+                                return provider.queryPresignedPutUrl(requesterId, contentType, keyStr, keyNum, contentLength);
+                            } else {
+                                return Mono.error(TurmsBusinessException.get(TurmsStatusCode.UNAUTHORIZED));
+                            }
+                        });
             } else {
                 throw TurmsBusinessException.get(TurmsStatusCode.FILE_TOO_LARGE);
             }
@@ -82,9 +80,70 @@ public class StorageService {
 
     public Mono<Boolean> deleteResource(@NotNull Long requesterId, @NotNull ContentType contentType, @Nullable String keyStr, @Nullable Long keyNum) {
         if (provider != null) {
-            return provider.deleteResource(requesterId, contentType, keyStr, keyNum);
+            return hasPermissionToDelete(requesterId, contentType, keyStr, keyNum)
+                    .flatMap(hasPermission -> {
+                        if (hasPermission) {
+                            return provider.deleteResource(requesterId, contentType, keyStr, keyNum);
+                        } else {
+                            return Mono.error(TurmsBusinessException.get(TurmsStatusCode.UNAUTHORIZED));
+                        }
+                    });
         } else {
             throw TurmsBusinessException.get(TurmsStatusCode.NOT_IMPLEMENTED);
+        }
+    }
+
+    private Mono<Boolean> hasPermissionToGet(@NotNull Long requesterId, @NotNull ContentType contentType, @Nullable String keyStr, @Nullable Long keyNum) {
+        switch (contentType) {
+            case PROFILE:
+            case GROUP_PROFILE:
+                return Mono.just(true);
+            case ATTACHMENT:
+                if (keyNum != null) {
+                    return messageService.isMessageSentToUserOrByUser(keyNum, requesterId);
+                } else {
+                    throw TurmsBusinessException.get(TurmsStatusCode.ILLEGAL_ARGUMENTS);
+                }
+            default:
+                throw new IllegalStateException("Unexpected value: " + contentType);
+        }
+    }
+
+    private Mono<Boolean> hasPermissionToPut(@NotNull Long requesterId, @NotNull ContentType contentType, @Nullable String keyStr, @Nullable Long keyNum) {
+        switch (contentType) {
+            case PROFILE:
+                return Mono.just(true);
+            case GROUP_PROFILE:
+                if (keyNum != null) {
+                    return groupMemberService.isOwnerOrManager(requesterId, keyNum);
+                } else {
+                    throw TurmsBusinessException.get(TurmsStatusCode.ILLEGAL_ARGUMENTS);
+                }
+            case ATTACHMENT:
+                if (keyNum != null) {
+                    return messageService.isMessageSentToUserOrByUser(keyNum, requesterId);
+                } else {
+                    throw TurmsBusinessException.get(TurmsStatusCode.ILLEGAL_ARGUMENTS);
+                }
+            default:
+                throw new IllegalStateException("Unexpected value: " + contentType);
+        }
+    }
+
+    private Mono<Boolean> hasPermissionToDelete(@NotNull Long requesterId, @NotNull ContentType contentType, @Nullable String keyStr, @Nullable Long keyNum) {
+        switch (contentType) {
+            case PROFILE:
+                return Mono.just(true);
+            case GROUP_PROFILE:
+                if (keyNum != null) {
+                    return groupMemberService.isOwnerOrManager(requesterId, keyNum);
+                } else {
+                    throw TurmsBusinessException.get(TurmsStatusCode.ILLEGAL_ARGUMENTS);
+                }
+            case ATTACHMENT:
+                return Mono.just(false);
+            default:
+                throw new IllegalStateException("Unexpected value: " + contentType);
         }
     }
 }
